Исправлены замечания ревьюера
