+++ src/test/java/TestQuestions.java
@@ -4,30 +4,39 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import pageobject.MainPage;
+
+import java.io.File;
+
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-public class TestQuestionsFirefox {
+public class
+TestQuestionsChrome {
     private  WebDriver driver;
     private String question;
     private String answer;
-    public TestQuestionsFirefox(String question, String answer){
+    public TestQuestionsChrome(String question, String answer){
         this.question = question;
         this.answer = answer;
     }
 
     @Before
     public void before(){
-        driver = new FirefoxDriver();
-
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.addArguments("--remote-allow-origins=*");
+        chromeOptions.setBinary(new File("/Applications/Google Chrome.app/Contents/MacOS/Google Chrome"));
+        driver = new ChromeDriver(chromeOptions);
+        //driver = new FirefoxDriver();
     }
     @Parameterized.Parameters
     public static Object[][] getOrderData() {
         return new Object[][] {
-                { "Сколько это стоит? И как оплатить?", "Сутки — 400 рублей. Оплата курьеру — наличными или картой."},
-                { "Хочу сразу несколько самокатов! Так можно?", "Пока что у нас так: один заказ — один самокат. Если хотите покататься с друзьями, можете просто сделать несколько заказов — один за другим."},
+               // { "Сколько это стоит? И как оплатить?", "Сутки — 400 рублей. Оплата курьеру — наличными или картой."},
+                //{ "Хочу сразу несколько самокатов! Так можно?", "Пока что у нас так: один заказ — один самокат. Если хотите покататься с друзьями, можете просто сделать несколько заказов — один за другим."},
                 {"Я жизу за МКАДом, привезёте?", "Да, обязательно. Всем самокатов! И Москве, и Московской области."}
         };
     }
@@ -35,7 +44,7 @@ public class TestQuestionsFirefox {
     public void after(){
         driver.quit();
     }
-    //Проверка вопросов о важном в Firefox
+    //Проверка вопросов о важном в Chrome
     @Test
     public void checkAnswer(){
         MainPage mainPage = new MainPage(driver);
